+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 18.07.2018
 
+* [Rust] Code of Kutulu - Handle explorer function (CodinGame)
 * [Rust] Code of Kutulu - Get closest wanderer (CodinGame)
 
 #### 17.07.2018